# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Filkx API",
        "description": "API documentation for the Tech Filkx e-commerce platform",
        "contact": {
            "email": "admin@filkx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.api.tech.filkx.com",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/api/admin/accounting/ledger": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "List paginated ledger entries (completed/cancelled orders as charges/refunds)",
                "operationId": "fdac701442de181e33a7359eefb2c79d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "charge",
                                "refund"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/accounting/invoices": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "List paginated invoices",
                "operationId": "22c57cd87f55b6efc888cca7e4ed22b5",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "paid",
                                "issued",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/accounting/invoices/{id}/pdf": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "Download an order invoice as PDF",
                "operationId": "bc278497b5587efcf76372d1ae4dbb04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file download"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/accounting/stats": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "Get accounting statistics (revenue, refunds, net revenue)",
                "operationId": "8fdcc83258effe9cde02e7f65ec8e68a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "totalRevenueMinor": {
                                                    "type": "integer",
                                                    "example": 100000
                                                },
                                                "totalRefundsMinor": {
                                                    "type": "integer",
                                                    "example": 5000
                                                },
                                                "netRevenueMinor": {
                                                    "type": "integer",
                                                    "example": 95000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/accounting/export": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "Export the ledger (completed/cancelled orders) as CSV",
                "operationId": "d435bcd9cfb3e3cf3e947a8ac7f63273",
                "responses": {
                    "200": {
                        "description": "CSV file download"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/billing/stats": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "Get billing statistics (revenue, active subscriptions, pending payments)",
                "operationId": "12370534e0b219d165a78f0af637eb53",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "revenueMinor": {
                                                    "type": "integer",
                                                    "example": 100000
                                                },
                                                "activeSubscriptions": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "pendingPaymentsCount": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/billing/payments/pending": {
            "get": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "List paginated pending payments (orders awaiting payment)",
                "operationId": "d326b822df408ab0db1d39fa79b831ce",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/billing/payments/{id}/confirm": {
            "post": {
                "tags": [
                    "Admin Accounting"
                ],
                "summary": "Approve or reject a pending payment",
                "operationId": "d7568f0a45da506cd8a80e14f47fcf22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approve"
                                ],
                                "properties": {
                                    "approve": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment approved or rejected"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}